+++ src/rvos/types.h
@@ -9,6 +9,6 @@ typedef unsigned long long uint64_t;
 /*
  * RISCV32: register is 32bits width
  */ 
-typedef uint32_t reg_t;
+typedef uint64_t reg_t;
 
 #endif /* __TYPES_H__ */
